updated button to place cards
