+++ webapp/src/lib/components/ButtonPlaceCard.svelte
@@ -3,14 +3,16 @@
 	export let accentColor: string = '#ff847c';
 </script>
 
-<div class="scale-75 w-[10rem] h-[4rem] relative flex items-center justify-center group">
-	<button
+<button
+	class="scale-75 w-[10rem] h-[4rem] relative flex items-center justify-center group"
+	on:click
+>
+	<div
 		class="p-2 h-[2rem] flex justify-center items-center bg-yellow group-hover:bg-purple transition-all z-20 absolute"
-		on:click
 	>
 		<p class="font-contrail text-sm group-hover:text-white transition-all">{text}</p>
-	</button>
-    <div
+	</div>
+	<div
 		class="invisible group-hover:visible flex justify-center items-center absolute w-[10rem] h-[15rem] border-dashed border-[.2rem] rounded-lg overflow-hidden transition-all"
 		style="border-color: {accentColor};"
 	></div>
@@ -21,4 +23,4 @@
 	<div
 		class="w-[5rem] h-[2rem] left-12 top-6 border-[0.4rem] z-0 rotate-[2deg] group-hover:rotate-[4deg] transition-all absolute"
 	/>
-</div>
+</button>
